# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Alchemist API
  description: |
    REST API for Alchemist - a self-hosted media transcoding pipeline.
    
    ## Authentication
    
    Most endpoints require authentication. Obtain a session token via `/auth/login` and then include it as:
    - **Cookie**: `alchemist_session=<token>`
    - **Header**: `Authorization: Bearer <token>`
    
    ## Rate Limiting
    
    - **Login attempts**: 10 per second
    - **Global API**: 120 requests per minute
    
    ## Versioning
    
    The API is available at both `/api/v1/*` (versioned) and `/api/*` (backwards-compatible alias).
  version: 1.0.0
  contact:
    name: Alchemist
    url: https://github.com/alchemist-dev/alchemist
  license:
    name: MIT
servers:
  - url: /api/v1
    description: Versioned API (recommended)
  - url: /api
    description: Backwards-compatible alias
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user
      description: Login with username and password to obtain a session token.
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              schema:
                type: string
                example: alchemist_session=abc123; HttpOnly; SameSite=Lax; Max-Age=2592000
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /settings/bundle:
    get:
      tags: [Settings]
      summary: Get entire configuration bundle
      operationId: getConfigBundle
      responses:
        '200':
          description: Full configuration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigBundle'
    put:
      tags: [Settings]
      summary: Update entire configuration bundle
      operationId: updateConfigBundle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfigBundle'
      responses:
        '200':
          description: Configuration updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigBundle'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
    cookieAuth:
      type: apiKey
      in: cookie
      name: alchemist_session
...
  schemas:
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
          example: ok
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
          format: password

    # --------------------------------------------------------------------------
    # Jobs
    # --------------------------------------------------------------------------
...
    ConfigBundle:
      type: object
      description: Complete configuration including all settings sections
      additionalProperties: true

    # --------------------------------------------------------------------------
    # Watch Directories
    # --------------------------------------------------------------------------
...
security:
  - bearerAuth: []
  - cookieAuth: []